+++ controllers/httproute_controller.go
@@ -21,7 +21,7 @@ import (
 	"github.com/go-logr/logr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/service-apis/api/v1alpha1"
+	"sigs.k8s.io/service-apis/apis/v1alpha1"
 )
 
 // HTTPRouteReconciler reconciles a HTTPRoute object
