+++ docs-src/concepts.md
@@ -229,7 +229,7 @@ TODO
 defined by the infrastructure provider. This resource represents a class of
 Gateways that can be instantiated.
 
-[gatewayclass-src]: https://github.com/kubernetes-sigs/service-apis/blob/master/api/v1alpha1/gatewayclass_types.go
+[gatewayclass-src]: https://github.com/kubernetes-sigs/service-apis/blob/master/apis/v1alpha1/gatewayclass_types.go
 
 > Note: this serves the same function as the [`networking.IngressClass` resource][ingress-class-api].
