+++ docs/concepts/index.html
@@ -1014,7 +1014,7 @@ be differences in behavior from the spec.</p>
 <h3 id="extension-points">Extension points</h3>
 <p>TODO</p>
 <h2 id="gatewayclass_1">GatewayClass</h2>
-<p><code>GatewayClass</code> (<a href="https://github.com/kubernetes-sigs/service-apis/blob/master/api/v1alpha1/gatewayclass_types.go">source code</a>) is cluster-scoped resource
+<p><code>GatewayClass</code> (<a href="https://github.com/kubernetes-sigs/service-apis/blob/master/apis/v1alpha1/gatewayclass_types.go">source code</a>) is cluster-scoped resource
 defined by the infrastructure provider. This resource represents a class of
 Gateways that can be instantiated.</p>
 <blockquote>
