+++ hack/update-proto.sh
@@ -24,7 +24,7 @@ go install k8s.io/code-generator/cmd/go-to-protobuf/protoc-gen-gogo
 # TODO(https://github.com/kubernetes-sigs/service-apis/issues/79): replace package name (k8s.io) with sigs.k8s.io
 go run k8s.io/code-generator/cmd/go-to-protobuf \
     --proto-import=$PWD/third_party/protobuf \
-    --packages sigs.k8s.io/service-apis/api/v1alpha1=k8s.io.service_apis.api.v1alpha1,-sigs.k8s.io/controller-runtime/pkg/scheme \
+    --packages sigs.k8s.io/service-apis/apis/v1alpha1=k8s.io.service_apis.api.v1alpha1,-sigs.k8s.io/controller-runtime/pkg/scheme \
     --drop-embedded-fields k8s.io/apimachinery/pkg/runtime.SchemeBuilder,sigs.k8s.io/controller-runtime/pkg/scheme \
     --apimachinery-packages -k8s.io/apimachinery/pkg/runtime/schema,-k8s.io/apimachinery/pkg/runtime,-k8s.io/apimachinery/pkg/apis/meta/v1,-k8s.io/api/core/v1 \
     --go-header-file hack/boilerplate.go.txt
